requirements.txt
matplotlib>=3.8
websocket-client
tk
numpy
scikit-learn
pandas
